sorting changed!

+++ apps/cart/views.py
@@ -6,7 +6,6 @@ from apps.cart.models import Cart
 
 
 def cart(request: WSGIRequest):
-
     if not request.user.is_authenticated:
         return redirect('login-page')
     context = {
@@ -17,10 +16,18 @@ def cart(request: WSGIRequest):
 
 @login_required(login_url='login-page')
 def create_cart(request: WSGIRequest, product_id: int):
-    Cart.objects.get_or_create(product_id=product_id, user=request.user)
+    if request.method == 'GET':
+        redirect('404-page')
+    quantity = request.GET.get('cart_quantity', 1)
+    obj,created = Cart.objects.get_or_create(product_id=product_id, user=request.user)
+    if obj.quantity != quantity:
+        obj.quantity = quantity
+        obj.save()
     return redirect(request.META['HTTP_REFERER'])
 
 
 def delete_cart(request: WSGIRequest, product_id: int) -> None:
+    if request.method == 'GET':
+        redirect('404-page')
     Cart.objects.filter(product_id=product_id).delete()
     return redirect(request.META['HTTP_REFERER'])
